--- src/pages/about/ui/index.ts
@@ -1,1 +1,0 @@
-export { AboutPage } from './about-page';
