--- src/pages/channel/ui/chat-item/index.ts
@@ -1,1 +1,0 @@
-export { ChatItem } from './chat-item';
