--- src/widgets/app-left-sidebar/model/channel.model.ts
@@ -1,7 +1,0 @@
-export class ChannelModel {
-  constructor(
-    public readonly id: string,
-    public readonly name: string,
-    public readonly creatorId: string,
-  ) {}
-}
