+++ src/widgets/app-left-sidebar/model/index.ts
@@ -1,1 +1,7 @@
-export { ChannelModel } from './channel.model';
+export class ChannelModel {
+  constructor(
+    public readonly id: string,
+    public readonly name: string,
+    public readonly creatorId: string,
+  ) {}
+}
