--- src/widgets/app-left-sidebar/ui/app-left-sidebar/index.ts
@@ -1,1 +1,0 @@
-export { AppLeftSidebar } from './app-left-sidebar';
