--- src/widgets/app-left-sidebar/ui/channel-list/index.ts
@@ -1,1 +1,0 @@
-export { ChannelList } from './channel-list';
