+++ src/widgets/app-left-sidebar/ui/channel-list/index.tsx
